Moved UniformLookupTableGenerator to header only
